+++ src/cargo.rs
@@ -23,6 +23,20 @@ impl CommandMessagesExt for CommandMessages {
                         .map(ToOwned::to_owned)
                         .map(Ok),
                     Ok(escargot::format::Message::CompilerMessage(e)) => {
+                        // We ignore the messages that are generated due to the use of `-Ctarget-feature`
+                        if e.message
+                            .message
+                            .contains("unstable feature specified for `-Ctarget-feature`")
+                        {
+                            return None;
+                        }
+                        // We also the "N warnings emitted" messages, because they are no longer accurate
+                        // since we ignore the "unstable feature specified for `-Ctarget-feature`" messages.
+                        if e.message.message.contains("warning emitted")
+                            || e.message.message.contains("warnings emitted")
+                        {
+                            return None;
+                        }
                         if let Some(rendered) = e.message.rendered {
                             eprint!("{rendered}");
                         }